+++ changelog.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- More internal docs
+- install-bx-module helper tests
+
+### Changed
+
+- Updated the helpers to be installed from the CDN
+
 ## [1.14.0] - 2025-06-24
 
 ### Added